Codingame.SourcePackager

Description

A small module to create monolithic sources from multiples files.

Documentation

createMonolithicSource :: FilePath -> IO String

Create a monolithic source by concatening a Haskell source file and all its local dependencies. The result is not just a string concatenation and source files are actually parsed and processed such as the returned source forms a valid Haskell content. Only the local content is incorported in the resulting source, whereas the external modules are simply kept as imported modules (eg. the source code of "System.IO" won’t be included. An internal module is a module whose source file can be found in the same hierachical directory as the initial file provided. Per instance, when creation a monolithic source from "src/Machin.hs", the source of an imported module named "Truc.Bidule" will be searched in "src/Truc". Last but not least, any function named "runMain" will be renamed into "main". It allows you to generate a valid program by only selecting a subset of your code base.

createMonolithicSourceWithMode :: ParseMode -> FilePath -> IO String

Create a monolithic source by concatening a Haskell source file and all its local dependencies. This function works the same way as createMonolithicSource but offer a way to customize the parse mode in order to deal with some parsing limitations. Indeed files are only parsed, not compiled and some things are beyond the parser capability, such as operator fixities. If you are defining you own operators in a file and use then into another, you need to declare them by overriding the parse mode:

let importedFixities = infixl_ 8 ["|>"]
    fixities' = importedFixities ++ fromMaybe [] (fixities defaultParseMode)
    parseMode = defaultParseMode{ fixities = Just fixities' }

source <- createMonolithicSourceWithMode parseMode "src/Golgoth.hs"

Unit testing

sourcePackagerTestGroup :: Test