Codingame.WebServices

Description

Give access to the Codingame web services API from your code. A typical usage is to test your code against another player into a test session:

credentials <- readCredentials "credentials.json"
let source = "main = putStrLn \"42\""
result <- play credentials "Coders Strike Back" source [UploadedCode, TrainingAi] Nothing

Complex requests

submit

Arguments

:: Credentials
A user credentials.

-> String
The challenge title.

-> String
The source to submit.

-> IO (Either SessionError ())
The error or nothing on success.

Submit a Haskell source for a given multiplayers challenge. The published source is pushed into the multiplayers arena and also made available in the IDE (after having refreshed the page if needed).

play

Arguments

:: Credentials
A user credentials.

-> String
The challenge title.

-> String
The source to submit (even if not used).

-> [AgentId]
The agents for the game.

-> Maybe String
Optional game options (including the random seed).

-> IO (Either SessionError GameResult)
The error or the game result on success.

Upload a Haskell source into the IDE and play a single game between two agents. Any of these agent could use the uploaded source, but it is not mandatory. However, even if not used, a source shall be provided (I don’t know why CG has bound together these two operations).

readCredentials

Arguments

:: String
Configuration file path.

-> IO Credentials
The returned credentials.

Utility function to load credentials from a JSON configuration file.

Elementary requests

wsLoginSiteV2

Arguments

:: String
Email.

-> String
Password.

-> Session Login

wsFindGamesPuzzleProgress

Arguments

:: Maybe Int
User ID.

-> Session [GamePuzzleProgress]

wsFindPastChallenges :: Session [Challenge]

wsFindByGameId

Arguments

:: Int
Replay ID.

-> Session GameResult

wsFindInformationByIdAndSaveGame

Arguments

:: Int
Replay ID.

-> Int
Used ID.

-> Session GameResult

wsGenerateSessionFromPuzzleIdV2

Arguments

:: Maybe Int
Used ID.

-> Int
Game ID.

-> Session TestSession

wsSubmit

Arguments

:: String
Test session handle.

-> Submit

-> Session Int

wsPlay

Arguments

:: String
Test session handle.

-> Play

-> Session GameResult

Request data

data Login

Constructors

Login

Fields

login_remembered :: Bool

login_authenticated :: Bool

login_user :: User

login_codinGamer :: Codingamer

login_languageId :: Int

login_userId :: Int

login_userEmail :: String

login_countryCode :: String

login_newFeatures :: [String]

login_enabledNotifications :: [String]

login_chatToken :: String

data User

Constructors

User

Fields

user_id :: Int

user_email :: String

user_languageId :: Int

user_valid :: Bool

data Codingamer

Constructors

Codingamer

Fields

codingamer_userId :: Int

codingamer_email :: Maybe String

codingamer_pseudo :: String

codingamer_countryId :: Maybe String

codingamer_publicHandle :: Maybe String

codingamer_privateHandle :: Maybe String

codingamer_registerOrigin :: Maybe String

codingamer_enable :: Bool

codingamer_userValid :: Bool

codingamer_schoolId :: Maybe Int

codingamer_rank :: Maybe Int

codingamer_avatar :: Int

codingamer_cover :: Maybe Int

codingamer_creationTime :: Maybe Int

codingamer_onlineSince :: Maybe Int

codingamer_hiddenFromFriendFinder :: Maybe Bool

codingamer_shareAllSolutions :: Maybe Bool

codingamer_tagline :: Maybe String

codingamer_level :: Maybe Int

codingamer_xp :: Maybe Int

data GamePuzzleProgress

Constructors

GamePuzzleProgress

Fields

progress_id :: Int

progress_level :: String

progress_rank :: Int

progress_thumbnailBinaryId :: Maybe Int

progress_previewBinaryId :: Maybe Int

progress_title :: String

progress_achievementScore :: Int

progress_validatorScore :: Int

progress_achievementCount :: Int

progress_doneAchievementCount :: Int

progress_totalScore :: Int

progress_unlocked :: Bool

progress_solvedCount :: Int

progress_optimCriteriaId :: Maybe String

progress_creationTime :: Int

data Challenge

Constructors

Challenge

Fields

challenge_title :: String

challenge_date :: Int

challenge_publicId :: String

challenge_cover1Id :: Maybe Int

challenge_logoId :: Int

challenge_descriptionJson :: Maybe String

challenge_endApplicationsDate :: Int

challenge_applicationsClossed :: Bool

data PublicGameResult

Constructors

PublicGameResult

Fields

publicGameResult_viewer :: String

publicGameResult_puzzleTitle :: [String]

publicGameResult_gameResult :: GameResult

publicGameResult_shareable :: Bool

publicGameResult_puzzleId :: Int

data GameResult

Constructors

GameResult

Fields

gameResult_gameId :: Int

gameResult_refereeInput :: String

gameResult_agents :: Maybe [Agent]

gameResult_frames :: [Frame]

gameResult_ranks :: [Int]

gameResult_scores :: [Double]

data Agent

Constructors

Agent

Fields

agent_agentId :: Int

agent_index :: Int

agent_codingamer :: Codingamer

agent_score :: Double

agent_valid :: Bool

data Frame

Constructors

Frame

Fields

frame_agentId :: Int

frame_gameInformation :: String

frame_view :: String

frame_summary :: Maybe String

frame_stdout :: Maybe String

frame_keyframe :: Maybe Bool

data TestSession

Constructors

TestSession

Fields

testSession_reportReady :: Bool

testSession_questionCount :: Int

testSession_globalRemainingTime :: Int

testSession_handle :: String

testSession_direct :: Bool

data Submit

Constructors

Submit

Fields

submit_code :: String

submit_programmingLanguageId :: String

data Play

Constructors

Play

Fields

play_code :: String

play_programmingLanguageId :: String

play_multi :: Multi

data Multi

Constructors

Multi

Fields

multi_agentsIds :: [AgentId]

multi_gameOptions :: Maybe String

The game options (the seed) if any (as provided by gameResult_refereeInput).

data AgentId

Constructors

UploadedCode

TrainingAi

UserId Int

data Credentials

Constructors

Credentials

Fields

credentials_email :: String

credentials_password :: String